pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "instanseg-brightfield"
version = "0.1.0"
description = "Train a brightfield nuclei-and-cells InstanSeg model using DAB membrane staining as self-supervised ground truth"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-image>=0.21",
    "opencv-python>=4.8",
    "openslide-python>=1.3",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "fastremap>=1.14",
    "instanseg-torch>=0.1.0",
]

[project.optional-dependencies]
train = [
    "instanseg-torch[full]>=0.1.0",
    "tensorboard>=2.14",
    "wandb>=0.15",
]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
